Enhance Payments page UI and refactor component
