+++ src/containers/MainLayout/styles.module.scss
@@ -13,7 +13,7 @@
 }
 
 
-@media screen and (max-width: 786px) {
+@media screen and (max-width: 768px) {
   .navbar{
     height: 80px;
     margin-top: 150px;
