+++ src/containers/main/PaymentsContainer/styles.module.scss
@@ -1,18 +1,38 @@
 @import "../../../assets/styles/mixins";
 
-@media screen and (max-width: 786px) {
-  .outer_container{
+@media screen and (max-width: 768px) {
+  .outer_container {
     width: 100vw;
   }
+
+  .payment_page {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    h2 {
+      color: lightblue;
+      padding-top: 10px;
+      text-align: center;
+    }
+  }
 }
 
-@media screen and (min-width: 786px) {
-  .outer_container{
+@media screen and (min-width: 768px) {
+  .outer_container {
     padding: 2.5em;
   }
-  
+
+  .payment_page {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    h2 {
+      color: lightblue;
+      padding-top: 10px;
+      text-align: center;
+    }
+  }
+
 }
-
-
-
-
