+++ src/pages/PaymentsPage/layout/checkout/FailCard/styles.module.scss
@@ -1,4 +1,4 @@
-@media screen and (max-width: 786px) {
+@media screen and (max-width: 768px) {
   .fail_wrapper{
     height: 100%;
     width: 80%;
@@ -33,7 +33,7 @@
 
 }
 
-@media screen and (min-width: 786px) {
+@media screen and (min-width: 768px) {
   .fail_wrapper{
     height: 100%;
     width: 100%;
