+++ src/pages/PaymentsPage/layout/checkout/StreamForm/styles.module.scss
@@ -1,15 +1,32 @@
-@import "../../../assets/styles/mixins";
+@import "../../../../../assets/styles/mixins";
 
-@media screen and (min-width: 786px) {
+@media screen and (min-width: 768px) {
+
+  .stream_form {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+    background-color: #2d3233;
+    border-radius: 10px;
+    color: lightblue;
+    font-size: 14px;
+    padding: 2em;
+    max-width: 800px;
+    align-content: center;
+    align-items: center;
+    justify-content: center;
+    justify-items: center;
+  }
 
   .label {
     font-weight: 500;
     font-size: 18px;
     line-height: 140%;
+    margin-top: 5px;
+    align-self: flex-start;
   }
 
   .input {
-    margin-top: -1em;
     padding: 0.7em;
     min-width: 15vw;
     background: #484e51;
@@ -24,7 +41,6 @@
     border-radius: 8px;
     border: 1px solid $dark-mode-lightblue;
     color: $dark-mode-lightblue;
-    margin-top: 1em;
     padding: 0.7em;
     width: 100%;
   }
@@ -41,14 +57,27 @@
     border: none;
     color: rgb(70, 68, 68);
     cursor: pointer;
+    margin-top: 5px;
   }
 }
 
-@media screen and (max-width: 786px) {
+@media screen and (max-width: 768px) {
   .stream_form {
     width: 95vw;
     padding: 1em;
-    padding-top: 60px;
+
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    background-color: #2d3233;
+    border-radius: 10px;
+    color: lightblue;
+    // font-size: 14px;
+    padding: 2em;
+    align-content: center;
+    align-items: center;
+    justify-content: center;
+    justify-items: center;
   }
 
   .close {
@@ -57,7 +86,7 @@
 
   .stream_form_container {
     width: 100%;
-    padding: 2em;
+    // padding: 2em;
   }
 
   .title {
@@ -86,7 +115,6 @@
     border-radius: 8px;
     border: 1px solid $dark-mode-lightblue;
     color: $dark-mode-lightblue;
-    margin-top: 1em;
     padding: 0.7em;
     width: 100%;
     height: 5vh;
@@ -97,7 +125,6 @@
     margin-top: 2em;
     height: 5vh;
     width: 80%;
-    margin-left: 10%;
     background-color: #678EB5;
     border-radius: 12px;
     transition: 0.5s;
