+++ src/pages/PaymentsPage/layout/checkout/StreamManagerItem/index.tsx
