+++ src/pages/PaymentsPage/styles.module.scss
@@ -2,7 +2,7 @@
 
 
 
-@media only screen and (max-width: 769px){
+@media only screen and (max-width: 768px){
 
   .content {  
     margin-left: 0;
@@ -18,7 +18,7 @@
   }
 }
 
-@media only screen and (min-width: 769px){
+@media only screen and (min-width: 768px){
 
   .content {
     margin-left: 12vw;
